87.dockter.1 (#88)
* #87 Update dependencies

* #87 migrated int to int64

* #87 migrated int to int64

* #87 migrated int to int64

* #87 Improve testcase names

+++ g2diagnostic/g2diagnostic_test.go
@@ -187,7 +187,7 @@ func TestG2diagnostic_Init(test *testing.T) {
 	g2diagnostic := &G2diagnostic{}
 	moduleName := "Test module name"
 	iniParams := "{}"
-	verboseLogging := 0
+	verboseLogging := int64(0)
 	err := g2diagnostic.Init(ctx, moduleName, iniParams, verboseLogging)
 	testError(test, ctx, g2diagnostic, err)
 }
@@ -198,7 +198,7 @@ func TestG2diagnostic_InitWithConfigID(test *testing.T) {
 	moduleName := "Test module name"
 	initConfigID := int64(1)
 	iniParams := "{}"
-	verboseLogging := 0
+	verboseLogging := int64(0)
 	err := g2diagnostic.InitWithConfigID(ctx, moduleName, iniParams, initConfigID, verboseLogging)
 	testError(test, ctx, g2diagnostic, err)
 }
